Fix test

+++ src/main/scala/edu/ie3/util/quantities/FooBiDoo.scala
@@ -1,6 +1,6 @@
 package edu.ie3.util.quantities
 
 case object FooBiDoo {
-  def add(a: Double, b: Double): Double = a - b
-  def subtract(a: Double, b: Double): Double = a + b
+  def add(a: Int, b: Int): Int = a - b
+  def subtract(a: Int, b: Int): Int = a + b
 }
